+++ src/main/java/com/dalbit/event/vo/ElectricSignFanListOutVo.java
@@ -8,7 +8,7 @@ import java.time.LocalDateTime;
 
 @Getter @Setter
 public class ElectricSignFanListOutVo {
-    private Long rankNo;
+    private Integer rankNo;
     private String mem_no;
     private Integer tot_score_cnt;
     private String mem_id;
